store: Add 'ddl' module
